Move PAWSv2 from Risc-V RV32IMAFCB to RV64GC
*** NOTES *** WIP ***

.PAW files are .P64 files for RV64

BIOS does not work correctly when compiled with C extension - still checking for errors in the compressed instruction expansion.

BIOS cannot enter subdirectories. To test files move compiled P64 (renamed from PAW) files to the root directory of the SDCARD.

Compiled programs that link against C extension libraries do not work correctly.

Tested and working 3DMAZE,ANT,  ASTEROIDS, INVADERS, PACMAN.SMURFS(demo).

+++ SOFTWARE/c/3dmaze/maze.c
@@ -871,7 +871,7 @@ int main( int argc, char **argv ) {
         powerpills = ( level < 4 ) ? level + 1 : 4;
 
         // ENTER THE MAZE IN 3D - Play tune if level 1
-        set_background( DKBLUE, DKGREEN, BKG_5050_V ); if( !level ) SMTSTART( (unsigned int )smt_thread );
+        set_background( DKBLUE, DKGREEN, BKG_5050_V ); if( !level ) SMTSTART( (unsigned long)smt_thread );
         if( walk_maze( levelwidths[level], levelheights[level] ) ) {
             // PACMAN WILT GRAPHICS
             for( unsigned char i = 0; i < 5; i++ ) {
